traffic_congestion/milkrun/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["tqtensor - <victortang212@gmail.com>"]
description = "Using AI to support the local community problem."
name = "milkrun"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
boto3 = "^1.26.50"
python = ">=3.8,<=3.11"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {allow-prereleases = true, version = "^23.1a1"}
isort = "^5.10.1"
pre-commit = "^2.20.0"
toml-sort = "^0.22.1"
